Epilogue Finished
- Removed redundant dialogue asset
- Added static variable to dialogue event tracker to determine epilogue outcome
- Modified jumping to main menu from epilogue to trigger shortly after dialogue closes
