--- Assets/Dialogue/EndingDialogue.asset
@@ -1,70 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e5e052e72a7c7004bbbf8170d8d437b4, type: 3}
-  m_Name: EndingDialogue
-  m_EditorClassIdentifier: 
-  dialogueNodes:
-  - id: 1
-    dialogueText: "1)You return to the small shack that you call your home, the land
-      around is desolate and your colleagues return from a long day in the mines. 
-      They have raw rings around their ankles where their shackles once were and
-      they look more tired than they ever have.  One man collapses as he tries to
-      open his door.  You rush over and attempt to help but he is already dead. 
-      You wonder if this was the right decision, if this wasn\u2019t your problem
-      then why wasn\u2019t it anyone else\u2019s?  Is it too late to change your
-      mind? "
-    hasNewActorLink: 0
-    linkActor: 
-    linkNodeId: 
-    responses: []
-  - id: 2
-    dialogueText: 'What do you do now? '
-    hasNewActorLink: 0
-    linkActor: 
-    linkNodeId: 
-    responses:
-    - responseText: "Head home. The cause isn\u2019t worth the risk."
-      responseId: 3
-      hasNewActorLink: 0
-      linkActor: 
-      linkNodeId: 
-    - responseText: Deliver the message. For a Free Vivat!
-      responseId: 4
-      hasNewActorLink: 0
-      linkActor: 
-      linkNodeId: 
-  - id: 3
-    dialogueText: "You discard the arrow head, not wanting to risk being caught with
-      a rebel symbol, and return to the small shack that you call your home, the
-      land around is desolate and your colleagues return from a long day in the mines. 
-      They have raw rings around their ankles where their shackles once were and
-      they look more tired than they ever have.  One man collapses as he tries to
-      open his door.  You rush over and attempt to help but he is already dead. 
-      You wonder if this was the right decision, if this wasn\u2019t your problem
-      then why wasn\u2019t it anyone else\u2019s?  Is it too late to change your
-      mind?"
-    hasNewActorLink: 0
-    linkActor: 
-    linkNodeId: 
-    responses: []
-  - id: 4
-    dialogueText: "You locate the rebel base and enter, some of the more armed commoners
-      confront you before you can get far and offer a careful welcome. Seeing their
-      guarded expressions, you bring out  the arrow head as you were instructed and
-      report that Riser Nine had sent you. Upon hearing your words, many commoners
-      like yourself now turned towards you with more hope than you\u2019ve ever seen
-      before in their eyes, as one of them questions you with disbelief, \u201CYou\u2019ve
-      escaped the dungeon?\u201D"
-    hasNewActorLink: 0
-    linkActor: 
-    linkNodeId: 
-    responses: []
