--- Assets/Dialogue/EndingDialogue.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ceac6b2e452469746817025f02d2d487
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
